Renamed `i18n-js` tech to `i18n`
