+++ test/techs/i18n/v2.test.js
@@ -3,10 +3,10 @@ var path = require('path'),
     serializeJS = require('serialize-javascript'),
     MockNode = require('mock-enb/lib/mock-node'),
     dropRequireCache = require('enb/lib/fs/drop-require-cache'),
-    Tech = require('../../../techs/i18n-js'),
+    Tech = require('../../../techs/i18n'),
     core = require('../../fixtures/bem-core-v3/common.blocks/i18n/i18n.i18n.js').i18n.i18n;
 
-describe('i18n-js v2', function () {
+describe('i18n v2 (bem-core)', function () {
     afterEach(function () {
         mock.restore();
     });
@@ -206,4 +206,3 @@ function build(keysets) {
 function serialize(js) {
     return 'module.exports = ' + serializeJS(js) + ';';
 }
-
